Demonstrate clustered session management with Tomcat
This version builds a war file that is run in standalone tomcat in a
docker container.

3 instances of the app run.

It also seeds the database to have a new tomcat_sessions table, as per
https://tomcat.apache.org/tomcat-8.5-doc/config/manager.html

Accessing http://localhost should return a session ID.

It also shows that the valueBound and valueUnbound lifecycle events are
happening as expected.

haproxy is load-balancing requests to the tomcat backends in a
round-robin fashion. There are no sticky sessions.

In this case, we are treating haproxy as an approximation of an AWS ELB.

We can destroy one of the tomcat containers (using `docker ps` and
`docker stop`). This failure isn't noticed by the client browser.
The session is still available, and contains the same context.

We can destroy all of the tomcat containers. This simulates an extended
outage, maybe due to an AWS AZ or region issue, or just a bad
deployment.

Once a tomcat has been running successfullyf for a minute or so,
the sessions are loaded and all expired, so that cleanup handled in a
valueUnbound implementation works fine.

Implementation Notes
====================

PersistentValve is required to ensure that sessions are flushed to the
DB immediately, and are then visible/loadable by other Tomcat nodes.
Otherwise non-sticky sessions don't work. Without it, sessions are
persisted to the DB depending on the configuration of `maxIdleSwap`,
`maxIdleBackup`, and related PersistentManager settings.

Concerns
========

There is a minor wrinkle. This does not handle concurrent requests under
the same session. We are only expecting a single HTML request in flight
at a time.

But concurrent requests for assets (which currently would require a
session) might cause problems.

Ditto HTTP2.